% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{calQCMetrics}
\alias{calQCMetrics}
\title{get QC metrics using gene counting matrix}
\usage{
calQCMetrics(scd)
}
\arguments{
\item{scd}{an SCData object containing count}
}
\value{
an SCData with updated QC metrics
}
\description{
get QC metrics using gene counting matrix
}
\details{
get QC metrics using gene counting matrix
the QC statistics added are
`number_of_genes`: number of genes detected
`total_count_per_cell`: sum of read number after UMI dedup
`non_mt_percent`: 1- percent of mitochondrial gene counts
mitochondrial genes are retrived by GO term GO:0005739
`non_ERCC_percent`: ratio of exon counts to ERCC counts
`non_ribo_percent`: 1- percent of ribosomal gene counts
ribosomal genes are retrived by GO term GO:0005840
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
scd = calQCMetrics(scd)
scd = detect_outlier(scd)

}
