% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_methods.R
\name{convert_geneid}
\alias{convert_geneid}
\title{convert the gene ids of a SCData object}
\usage{
convert_geneid(scd, returns = "external_gene_name", all = TRUE)
}
\arguments{
\item{scd}{an SCData object}

\item{returns}{the gene id which is set as return. default to be `external_gene_name`
A possible list of attributes can be retrieved using the
function \code{listAttributes} from \code{biomaRt} package. the commonly used
id types are `external_gene_name`, `ensembl_gene_id` or `entrezgene`.}

\item{all}{logic. for genes that cannot covert to new gene id, keep them with the old
id or delete them. the default is keep them.}
}
\value{
scd with converted id
}
\description{
convert the gene ids of a SCData object
}
\details{
convert the gene id of all datas in SCData object
}
\examples{
#TODO

}
