% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts<-}
\alias{counts}
\alias{counts.SCData}
\alias{counts}
\alias{counts,SCData-method}
\alias{counts<-,SCData,matrix-method}
\alias{counts,SCData-method}
\alias{counts<-,SCData,matrix-method}
\title{Accessors for the 'counts' element of an SCData object.}
\usage{
counts(object)

counts(object) <- value

counts.SCData(object)

\S4method{counts}{SCData}(object)

\S4method{counts}{SCData,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCData} object.}

\item{value}{an integer matrix}
}
\value{
a matrix of count data

the raw count
}
\description{
The counts element holds the count data as a matrix of non-negative integer
count values, one row for each feature (gene, exon, region, etc), and one
column for each cell. It is an element of the assayData slot of the SCData
object.
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
counts(scd)

}
\author{
Luyi Tian
}
