% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_workflow.R
\name{create_scd_by_dir}
\alias{create_scd_by_dir}
\title{create SCData object from data folder generated by preprocessing step}
\usage{
create_scd_by_dir(datadir, org = NULL, gene_id = NULL, pheno_data = NULL)
}
\arguments{
\item{datadir}{the directory that contains all the data and `stat` subfolder.}

\item{org}{the organism of the data. List of possible names can be retrieved using the function 
`listDatasets`from `biomaRt` package. (i.e `mmusculus_gene_ensembl` or `hsapiens_gene_ensembl`)}

\item{gene_id}{gene id type of the data A possible list of ids can be retrieved using the function `listAttributes` from `biomaRt` package. 
the commonly used id types are `external_gene_name`, `ensembl_gene_id` or `entrezgene`}

\item{pheno_data}{the external phenotype data that linked to each single cell. this should be an \code{AnnotatedDataFrame} object}
}
\value{
an SCData object
}
\description{
after we run \code{sc_gene_counting} and finish the preprocessing step. \code{create_scd_by_dir}
can be used to generate the SCData obeject from the folder that contains gene counting matrix and QC statistics.
}
\details{
after we run \code{sc_gene_counting} and finish the preprocessing step. \code{create_scd_by_dir}
can be used to generate the SCData obeject from the folder that contains gene counting matrix and QC statistics.
}
