% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_workflow.R
\name{createReport}
\alias{createReport}
\title{createReport}
\usage{
createReport(sample_name, outdir, r1, r2 = NULL, outfq, read_structure,
  filter_settings, align_bam, genome_index, map_bam, exon_anno, stnd, fix_chr,
  barcode_anno, max_mis, UMI_cor, gene_fl, organism, gene_id_type)
}
\arguments{
\item{sample_name}{sample name}

\item{outdir}{output folder}

\item{r1}{file path of read1}

\item{r2}{file path of read2 default to be NULL}

\item{outfq}{file path of the output of \code{sc_trim_barcode}}

\item{read_structure}{a list contains read structure configuration. for more help see `?sc_trim_barcode`}

\item{filter_settings}{a list contains read filter settings for more help see `?sc_trim_barcode`}

\item{align_bam}{the aligned bam file}

\item{genome_index}{genome index used for alignment}

\item{map_bam}{the mapped bam file}

\item{exon_anno}{the gff exon annotation used. can have multiple files}

\item{stnd}{whether to perform strand specific mapping}

\item{fix_chr}{add `chr` to chromosome names, fix inconsistant names.}

\item{barcode_anno}{cell barcode annotation file path.}

\item{max_mis}{maximum mismatch allowed in barcode. default to be 1}

\item{UMI_cor}{correct UMI sequence error: 0 means no correction, 1 means simple correction and merge UMI with distance 1.}

\item{gene_fl}{whether to remove low abundant gene count. low abundant is defined as only one copy of one UMI for this gene}

\item{organism}{the organism of the data. List of possible names can be retrieved using the function 
`listDatasets`from `biomaRt` package. (i.e `mmusculus_gene_ensembl` or `hsapiens_gene_ensembl`)}

\item{gene_id_type}{gene id type of the data A possible list of ids can be retrieved using the function `listAttributes` from `biomaRt` package. 
the commonly used id types are `external_gene_name`, `ensembl_gene_id` or `entrezgene`}
}
\description{
create html report using data generated by proprocessing step.
}
