% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCData_methods.R
\name{demultiplex_info.sce}
\alias{demultiplex_info.sce}
\title{demultiplex_info}
\usage{
demultiplex_info.sce(object)
}
\arguments{
\item{object}{An \code{SingleCellExperiment} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
A DataFrame of cell barcode demultiplx results.
}
\description{
Get or set cell barcode demultiplx results in a SingleCellExperiment object
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
QCMetrics(scd)

}
\author{
Luyi Tian
}
