% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{DimReducedExpr}
\alias{DimReducedExpr}
\alias{DimReducedExpr<-}
\alias{DimReducedExpr,SCData-method}
\alias{DimReducedExpr}
\alias{DimReducedExpr<-,SCData-method}
\alias{DimReducedExpr}
\title{Get or set \code{reducedExprDimension} from an SCData object}
\usage{
DimReducedExpr(object)

DimReducedExpr(object) <- value

\S4method{DimReducedExpr}{SCData}(object)

\S4method{DimReducedExpr}{SCData}(object) <- value
}
\arguments{
\item{object}{An \code{\link{SCData}} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
A matrix of reduced-dimension coordinates for gene
expression of single cell.
}
\description{
Get or set \code{reducedExprDimension} from an SCData object
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
DimReducedExpr(scd)

}
\author{
Luyi Tian
}
