% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{DimReduceExpr}
\alias{DimReduceExpr}
\alias{DimReduceExpr<-}
\alias{DimReduceExpr}
\alias{DimReduceExpr.SCData}
\alias{DimReduceExpr}
\alias{DimReduceExpr<-}
\alias{DimReduceExpr}
\title{Get or set \code{reducedExprDimension} from an SCData object}
\usage{
DimReduceExpr(object)

DimReduceExpr(object) <- value

DimReduceExpr.SCData(object)

\S4method{DimReduceExpr}{SCData}(object)

\S4method{DimReduceExpr}{SCData}(object) <- value
}
\arguments{
\item{object}{An \code{\link{SCData}} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
A matrix of reduced-dimension coordinates for gene
expression of single cell.
}
\description{
Get or set \code{reducedExprDimension} from an SCData object
}
\author{
Luyi Tian
}
