% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{.qq_outliers_robust}
\alias{.qq_outliers_robust}
\title{Detect outliers based on robust linear regression of QQ plot}
\usage{
.qq_outliers_robust(x, df, conf)
}
\arguments{
\item{x}{a vector of mahalanobis distance}

\item{df}{degree of freedom for chi-square distribution}

\item{conf}{confidence for linear regression}
}
\value{
cell names of outliers
}
\description{
Detect outliers based on robust linear regression of QQ plot
}
