% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCE_methods.R
\docType{methods}
\name{feature_info}
\alias{feature_info}
\alias{feature_info<-}
\alias{feature_info.sce}
\alias{feature_info,SingleCellExperiment-method}
\alias{feature_info<-,SingleCellExperiment-method}
\title{Get or set \code{feature_info} from a SingleCellExperiment object}
\usage{
feature_info(object)

feature_info(object) <- value

feature_info.sce(object)

\S4method{feature_info}{SingleCellExperiment}(object)

\S4method{feature_info}{SingleCellExperiment}(object) <- value
}
\arguments{
\item{object}{A \code{\link{SingleCellExperiment}} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
a dataframe of feature info for scATAC-seq data

A DataFrame of feature information
}
\description{
Get or set \code{feature_info} from a SingleCellExperiment object
}
\author{
Shani Amarasinghe
}
