% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCE_methods.R
\docType{methods}
\name{feature_type}
\alias{feature_type}
\alias{feature_type<-}
\alias{feature_type.sce}
\alias{feature_type,SingleCellExperiment-method}
\alias{feature_type<-,SingleCellExperiment-method}
\title{Get or set \code{feature_type} from a SingleCellExperiment object}
\usage{
feature_type(object)

feature_type(object) <- value

feature_type.sce(object)

\S4method{feature_type}{SingleCellExperiment}(object)

\S4method{feature_type}{SingleCellExperiment}(object) <- value
}
\arguments{
\item{object}{A \code{\link{SingleCellExperiment}} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
the feature type used in feature counting for scATAC-Seq data

A string representing the feature type
}
\description{
Get or set \code{feature_type} from a SingleCellExperiment object
}
\author{
Shani Amarasinghe
}
