% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{fpkm}
\alias{fpkm}
\alias{fpkm,SCData-method}
\alias{fpkm.SCData}
\alias{fpkm<-}
\alias{fpkm<-,SCData,matrix-method}
\title{Accessors for the 'fpkm' (fragments per kilobase of exon per million reads mapped) element of an SCData object.}
\usage{
fpkm(object)

fpkm(object) <- value

\S4method{fpkm}{SCData}(object)

\S4method{fpkm}{SCData,matrix}(object)<-value

\S4method{fpkm}{SCData}(object)

\S4method{fpkm}{SCData,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCData} object.}

\item{value}{a matrix of class \code{"numeric"}}
}
\value{
a matrix of FPKM values
}
\description{
The \code{fpkm} element of the arrayData slot in an SCData object holds
a matrix containing fragments per kilobase of exon per million reads mapped
(FPKM) values. It has the same dimensions as the 'exprs' and 'counts'
elements, which hold the transformed expression data and count data,
respectively.
}
\examples{

}
\author{
Luyi Tian
}

