% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{gene_id_type}
\alias{gene_id_type}
\alias{gene_id_type<-}
\alias{gene_id_type}
\alias{gene_id_type.SCData}
\alias{gene_id_type,SCData-method}
\alias{gene_id_type}
\alias{gene_id_type<-,SCData-method}
\alias{gene_id_type}
\title{Get or set \code{gene_id_type} from an SCData object}
\usage{
gene_id_type(object)

gene_id_type(object) <- value

gene_id_type.SCData(object)

\S4method{gene_id_type}{SCData}(object)

\S4method{gene_id_type}{SCData}(object) <- value
}
\arguments{
\item{object}{An \code{\link{SCData}} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
gene id type string
}
\description{
Get or set \code{gene_id_type} from an SCData object
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
gene_id_type(scd)

}
\author{
Luyi Tian
}
