% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_bam_tagging.R
\name{generate_table_graph}
\alias{generate_table_graph}
\title{importFrom ggplot2 ggplot geom_col aes ggsave}
\usage{
generate_table_graph(logs_and_stats_folder)
}
\description{
importFrom ggplot2 ggplot geom_col aes ggsave
}
