% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_create_fragments.R
\name{get_chromosomes}
\alias{get_chromosomes}
\title{Get Chromosomes}
\usage{
get_chromosomes(bam, keep_contigs = "^chr")
}
\arguments{
\item{bam}{file path to the bam file to get data from}

\item{keep_contigs}{regular expression used with grepl to filter reference names}
}
\value{
a named list where element names are chromosomes reference names and elements are integer lengths
}
\description{
Gets a list of NamedList of chromosomes and the reference length
acquired through the bam index file.
}
