% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ercc_anno}
\alias{get_ercc_anno}
\title{Get ERCC annotation table}
\usage{
get_ercc_anno()
}
\value{
data.frame containing ERCC annotation
}
\description{
Helper function to retrieve ERCC annotation as a dataframe in SAF format
}
\examples{
ercc_anno <- get_ercc_anno()

}
