% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_str.R
\name{get_read_str}
\alias{get_read_str}
\title{Get read structure for particular scRNA-seq protocol}
\usage{
get_read_str(protocol)
}
\arguments{
\item{protocol}{name of the protocol}
}
\value{
list of UMI and Barcode locations for use in other scPipe functions
}
\description{
The supported protocols are:
* CelSeq
* CelSeq2
* DropSeq
* 10x
If you know the structure of a specific protocol and would like it supported,
please leave a issue post at www.github.com/luyitian/scPipe.
}
\examples{
get_read_str("celseq")
}
