% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_methods.R
\name{getGenesByGO}
\alias{getGenesByGO}
\title{Get genes related to certain GO terms from biomart database}
\usage{
getGenesByGO(returns = "ensembl_gene_id",
  dataset = "mmusculus_gene_ensembl", go = NULL)
}
\arguments{
\item{returns}{the gene id which is set as return. default to be ensembl id
A possible list of attributes can be retrieved using the
function \code{listAttributes} from \code{biomaRt} package. the commonly used
id types are `external_gene_name`, `ensembl_gene_id` or `entrezgene`.}

\item{dataset}{Dataset you want to use. List of possible datasets can be
retrieved using the function \code{listDatasets} from \code{biomaRt} package.}

\item{go}{a vector of GO terms}
}
\value{
a vector of gene ids.
}
\description{
Get genes related to certain GO terms from biomart database
}
\details{
Get genes related to certain GO terms from biomart database
}
\examples{
#TODO

}
