% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_setup.R
\docType{data}
\name{list_of_packages}
\alias{list_of_packages}
\title{Set up the background forscATAC-Seq pre-processing
sc_atac_setup}
\format{
An object of class \code{character} of length 14.
}
\usage{
list_of_packages
}
\description{
Set up the background forscATAC-Seq pre-processing
sc_atac_setup
}
\examples{
\dontrun{
setpaths(macs2path,bedToolsPath)

}

}
\keyword{datasets}
