% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_demultiplex}
\alias{plot_demultiplex}
\title{plot_demultiplex}
\usage{
plot_demultiplex(sce)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}
}
\value{
a bar chart
}
\description{
plot cell barcode demultiplex result for the experiment.
It will include the percentage of reads that match the cell barcode.
For the reads that does not match any barcode, it will gives the percentage of
reads aligned and mapped to exon. If there is high proportion of reads
that do not match any barcode but mapped to exon, it might indicatea a failure in cell
barcode demultiplexing
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
sce = SingleCellExperiment(assays = list(counts =as.matrix(sc_sample_data)))
organism(sce) = "mmusculus_gene_ensembl"
gene_id_type(sce) = "ensembl_gene_id"
QC_metrics(sce) = sc_sample_qc
demultiplex_info(sce) = cell_barcode_matching
UMI_dup_info(sce) = UMI_duplication

plot_demultiplex(sce)

}
