% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_mapping}
\alias{plot_mapping}
\title{plot mapping statistics for SCData object.}
\usage{
plot_mapping(scd, sel_col = NULL, percentage = FALSE, dataname = "")
}
\arguments{
\item{scd}{an SCData object}

\item{sel_col}{a vector of column names, indicate the columns to use for plot.
by default it will be the mapping result.}

\item{percentage}{logic. whether to convert the number of reads to percentage}

\item{dataname}{the name of this dataset which appears in the plot title}
}
\value{
a ggplot2 object
}
\description{
plot mapping statistics for SCData object.
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
scd = calQCMetrics(scd)
scd = detect_outlier(scd)
plot_mapping(scd,percentage=TRUE,dataname="sc_sample")

}
