% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_UMI_dup}
\alias{plot_UMI_dup}
\title{plot_UMI_dup}
\usage{
plot_UMI_dup(sce, log10_x = TRUE)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}

\item{log10_x}{whether to use log10 scale for x axis}
}
\value{
a line chart of the UMI duplication distributions
}
\description{
plot UMI duplication number distributions for
each UMI sequence in each genes.
}
\examples{
# TODO
}
