% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotQC_pair}
\alias{plotQC_pair}
\title{plot QC statistics for \code{SingleCellExperiment} object}
\usage{
plotQC_pair(sce, sel_col = NULL)
}
\arguments{
\item{sce}{an \code{SingleCellExperiment} object}

\item{sel_col}{a vector of column names, indicate the columns to use for plot.
by default it will be the statistics generated by `calQC_metrics()`}
}
\value{
a ggplot2 object
}
\description{
plot QC statistics for \code{SingleCellExperiment} object
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
sce = new\\code{SingleCellExperiment}(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
sce = calQC_metrics(scd)
sce = detect_outlier(scd)
plotQC_pair(scd)

}
