% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{QC_metrics}
\alias{QC_metrics}
\alias{QC_metrics,SCData-method}
\alias{QC_metrics<-}
\title{Get or set cell names from an SCData object}
\usage{
QC_metrics(object)

QC_metrics(object) <- value

QC_metrics(object)

\S4method{QC_metrics}{SCData}(object)

\S4method{QC_metrics}{SCData}(object) <- value
}
\arguments{
\item{object}{An \code{\link{SCData}} object.}
}
\value{
A matrix of quality control metrics.
}
\description{
Get or set cell names from an SCData object
}
\examples{
TODO


}

