% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{QCMetrics}
\alias{QCMetrics}
\alias{QCMetrics<-}
\alias{QCMetrics}
\alias{QCMetrics.SCData}
\alias{QCMetrics}
\alias{QCMetrics<-}
\alias{QCMetrics}
\title{Get or set quality control metrics from an SCData object}
\usage{
QCMetrics(object)

QCMetrics(object) <- value

QCMetrics.SCData(object)

\S4method{QCMetrics}{SCData}(object)

\S4method{QCMetrics}{SCData}(object) <- value
}
\arguments{
\item{object}{An \code{\link{SCData}} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
A \code{AnnotatedDataFrame} of quality control metrics.
}
\description{
Get or set quality control metrics from an SCData object
}
\author{
Luyi Tian
}
