% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_workflow.R
\name{run_scPipe}
\alias{run_scPipe}
\title{run_scPipe}
\usage{
run_scPipe(sample_name, outdir, r1, r2 = NULL, read_structure,
  filter_settings = list(rmlow = TRUE, rmN = TRUE, minq = 20, numbq = 2),
  genome_index, exon_anno, fix_chr = FALSE, barcode_anno = NULL,
  max_mis = 1, has_UMI = TRUE, UMI_cor = 1, gene_fl = FALSE,
  organism = "NA", gene_id_type = "NA", nthreads = 1, report = TRUE)
}
\arguments{
\item{sample_name}{sample names, the name should not have any space.}

\item{outdir}{the output folder}

\item{r1}{file path of read1}

\item{r2}{file path of read2}

\item{read_structure}{a list contains read structure configuration. for more help see `?sc_trim_barcode`}

\item{filter_settings}{a list contains read filter settings for more help see `?sc_trim_barcode`}

\item{genome_index}{genome index used for \code{Rsubread::align}}

\item{exon_anno}{a vector of gff exon annotation file paths.}

\item{fix_chr}{whether add `chr` to chromosome names, fix inconsistant names between different annotations.}

\item{barcode_anno}{file path for cell barcode annotation.}

\item{max_mis}{maximum mismatch allowed in barcode. default to be 1}

\item{has_UMI}{whether the protocol has UMI, default to be TRUE}

\item{UMI_cor}{correct UMI sequence error: 0 means no correction, 1 means simple correction and merge UMI with distance 1.}

\item{gene_fl}{whether to remove low abundant gene count. low abundant is defined as only one copy of one UMI for this gene}

\item{organism}{the organism of the data. List of possible names can be retrieved using the function 
`listDatasets`from `biomaRt` package. (i.e `mmusculus_gene_ensembl` or `hsapiens_gene_ensembl`)}

\item{gene_id_type}{gene id type of the data A possible list of ids can be retrieved using the function `listAttributes` from `biomaRt` package. 
the commonly used id types are `external_gene_name`, `ensembl_gene_id` or `entrezgene`}

\item{nthreads}{number of threads used}

\item{report}{whether to generate the html report}
}
\value{
an SCData object
}
\description{
run_scPipe
}
\examples{
\dontrun{
 # geneate scd from fastq files
 scd = run_scPipe(sample_name="sample_001",
       outdir="output_dir_of_scPipe",
       r1="read1.fq",
       r2="read2.fq",
       read_structure=list(bs1=-1, bl1=2, bs2=6, bl2=8, us=0, ul=6),
       filter_settings=list(rmlow=TRUE, rmN=TRUE, minq=20, numbq=2),
       genome_index="mouse.index",
       exon_anno="exon_anno.gff3",
       has_UMI=TRUE,
       stnd=TRUE,
       fix_chr=FALSE,
       barcode_anno="cell_barcode.csv",
       max_mis=1,
       UMI_cor=1,
       gene_fl=FALSE,
       nthreads=1,
       report=TRUE,
       organism="mmusculus_gene_ensembl",
       gene_id_type="ensembl_gene_id")
}
# or directly from the gene count and quality control matrix:
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
}
