% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_bam_tagging.R
\name{sc_atac_bam_tagging}
\alias{sc_atac_bam_tagging}
\title{BAM tagging}
\usage{
sc_atac_bam_tagging(
  inbam,
  output_folder = NULL,
  bc_length = NULL,
  bam_tags = list(bc = "CB", mb = "OX"),
  nthreads = 1
)
}
\arguments{
\item{inbam}{The input BAM file}

\item{output_folder}{The path of the output folder}

\item{bc_length}{The length of the cellular barcodes}

\item{bam_tags}{The BAM tags}

\item{nthreads}{The number of threads}
}
\value{
file path of the resultant demmultiplexed BAM file.
}
\description{
Demultiplexes the reads
}
\details{
sc_atac_bam_tagging()
}
\examples{
inbam <- system.file("extdata", "tiny_sample.bam", package="scPipe")
out <- tempdir()

out_bam <- sc_atac_bam_tagging(
   inbam = inbam,
   output_folder = out,
   nthreads = 6)

}
