% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_call_peaks.R
\name{sc_atac_call_peaks}
\alias{sc_atac_call_peaks}
\title{sc_atac_call_peaks()}
\usage{
sc_atac_call_peaks(inbam, output_folder = "")
}
\value{

}
\description{
sc_atac_call_peaks()
}
\examples{
\dontrun{


}

}
