% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_cell_calling.R
\name{sc_atac_cellranger_cell_calling}
\alias{sc_atac_cellranger_cell_calling}
\title{cellranger cell calling}
\usage{
sc_atac_cellranger_cell_calling(mat, cell_qc_metrics_file, genome_size)
}
\arguments{
\item{mat}{The input matrix}

\item{cell_qc_metrics_file}{A file containing qc statistics for each cell}

\item{genome_size}{The size of the genome}
}
\description{
use the cellranger cell calling algorithm
}
