% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_create_cell_qc_metrics.R
\name{sc_atac_create_cell_qc_metrics}
\alias{sc_atac_create_cell_qc_metrics}
\title{generating a file useful for producing the qc plots}
\usage{
sc_atac_create_cell_qc_metrics(
  frags_file,
  peaks_file,
  promoters_file,
  tss_file,
  enhs_file,
  output_folder
)
}
\arguments{
\item{frags_file}{The fragment file}

\item{peaks_file}{The peak file}

\item{promoters_file}{The path of the promoter annotation file}

\item{tss_file}{The path of the tss annotation file}

\item{enhs_file}{The path of the enhs annotation file}

\item{output_folder}{The path of the output folder for resultant files}
}
\value{
Nothing (Invisible 'NULL')
}
\description{
uses the peak file and annotation files for features
}
