% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_create_fragments.R
\name{sc_atac_create_fragments}
\alias{sc_atac_create_fragments}
\title{DEPRECIATED. Generating the popular fragments for scATAC-Seq data using sinto}
\usage{
sc_atac_create_fragments(inbam, output_folder = "")
}
\arguments{
\item{inbam}{The tagged, sorted and duplicate-free input BAM file}

\item{output_folder}{The path of the output folder}
}
\description{
DEPRECIATED. Use sc_atac_fragments instead. Takes in a tagged and sorted BAM file and outputs the associated fragments in a .bed file
}
