% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_create_qc_per_bc_file.R
\name{sc_atac_create_qc_per_bc_file}
\alias{sc_atac_create_qc_per_bc_file}
\title{generating a file useful for producing the qc plots}
\usage{
sc_atac_create_qc_per_bc_file(
  inbam,
  frags_file,
  peaks_file,
  promoters_file,
  tss_file,
  enhs_file,
  output_folder,
  lower = NULL
)
}
\arguments{
\item{inbam}{The input bam file}

\item{frags_file}{The fragment file}

\item{peaks_file}{The peak file}

\item{promoters_file}{The path of the promoter annotation file}

\item{tss_file}{The path of the tss annotation file}

\item{enhs_file}{The path of the enhs annotation file}

\item{output_folder}{A string indicating the path of the output folder}

\item{lower}{the lower threshold for the data if using the \code{emptydrops} function for cell calling.}

\item{min_uniq_frags}{The minimum number of required unique fragments required for a cell (used for \code{filter} cell calling)}

\item{max_uniq_frags}{The maximum number of required unique fragments required for a cell (used for \code{filter} cell calling)}

\item{min_frac_peak}{The minimum proportion of fragments in a cell to overlap with a peak (used for \code{filter} cell calling)}

\item{min_frac_tss}{The minimum proportion of fragments in a cell to overlap with a tss (used for \code{filter} cell calling)}

\item{min_frac_enhancer}{The minimum proportion of fragments in a cell to overlap with a enhancer sequence (used for \code{filter} cell calling)}

\item{min_frac_promoter}{The minimum proportion of fragments in a cell to overlap with a promoter sequence (used for \code{filter} cell calling)}

\item{max_frac_mito}{The maximum proportion of fragments in a cell that are mitochondrial (used for \code{filter} cell calling)}
}
\description{
uses the peak file and annotation files for features
}
