% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_create_report.R
\name{sc_atac_create_report}
\alias{sc_atac_create_report}
\title{HTML report generation}
\usage{
sc_atac_create_report(
  input_folder,
  output_folder = NULL,
  sample_name = NULL,
  organism = NULL,
  feature_type = NULL
)
}
\arguments{
\item{input_folder}{The path of the folder produced by the pipeline}

\item{output_folder}{The path of the output folder to store the HTML report in}

\item{sample_name}{A string indicating the name of the sample}

\item{organism}{A string indicating the name of the organism being analysed}

\item{feature_type}{A string indicating the type of the feature (`genome_bin` or `peak`)}
}
\description{
Generates a HTML report using the output folder produced by the pipeline
}
