% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_cell_calling.R
\name{sc_atac_emptydrops_cell_calling}
\alias{sc_atac_emptydrops_cell_calling}
\title{empty drops cell calling}
\usage{
sc_atac_emptydrops_cell_calling(mat, output_folder, lower = NULL)
}
\arguments{
\item{mat}{The input matrix}

\item{output_folder}{The path of the output folder}

\item{lower}{Lower}
}
\description{
The empty drops cell calling method
}
