% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_feature_counting.R
\name{sc_atac_generate_per_ber_pc_file}
\alias{sc_atac_generate_per_ber_pc_file}
\alias{sc_atac_create_qc_per_bc_file}
\title{generating a file useful for producing the qc plots}
\usage{
sc_atac_create_qc_per_bc_file(
  inbam,
  frags_file,
  peaks_file,
  promoters_file,
  tss_file,
  enhs_file,
  output_folder
)
}
\arguments{
\item{inbam}{The input bam file}

\item{frags_file}{The fragment file}

\item{peaks_file}{The peak file}

\item{promoters_file}{The path of the promoter annotation file}

\item{tss_file}{The path of the tss annotation file}

\item{enhs_file}{The path of the enhs annotation file}

\item{output_folder}{}
}
\description{
uses the peak file and annotation files for features
}
