% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_feature_counting.R
\name{sc_atac_generate_qc_plots}
\alias{sc_atac_generate_qc_plots}
\title{generating qc plots}
\usage{
sc_atac_generate_qc_plots(
  frags_file,
  peaks_file,
  qc_per_bc_file,
  output_folder
)
}
\arguments{
\item{output_folder}{}
}
\description{
qc plots are produced using fragment file and peak file for features
}
