% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_peak_calling.R
\name{sc_atac_peak_calling}
\alias{sc_atac_peak_calling}
\title{sc_atac_peak_calling()}
\usage{
sc_atac_peak_calling(
  inbam,
  ref = NULL,
  output_folder = NULL,
  genome_size = NULL
)
}
\value{

}
\description{
sc_atac_peak_calling()
}
\examples{
\dontrun{


}

}
