% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_pipeline.R
\name{sc_atac_pipeline_quick_test}
\alias{sc_atac_pipeline_quick_test}
\title{A function that tests the pipeline on a small test sample (without duplicate removal)}
\usage{
sc_atac_pipeline_quick_test()
}
\value{
None (invisible `NULL`)
}
\description{
A function that tests the pipeline on a small test sample (without duplicate removal)
}
