% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_subfunctions.R
\name{sc_atac_plot_cells_per_feature}
\alias{sc_atac_plot_cells_per_feature}
\title{A histogram of the log-number of cells per feature}
\usage{
sc_atac_plot_cells_per_feature(sce)
}
\arguments{
\item{sce}{The SingleExperimentObject produced by the sc_atac_create_sce function at the end of the pipeline}
}
\value{
returns NULL
}
\description{
A histogram of the log-number of cells per feature
}
