% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_subfunctions.R
\name{sc_atac_plot_features_per_cell_ordered}
\alias{sc_atac_plot_features_per_cell_ordered}
\title{Plot showing the number of features per cell in ascending order}
\usage{
sc_atac_plot_features_per_cell_ordered(sce)
}
\arguments{
\item{sce}{The SingleExperimentObject produced by the sc_atac_create_sce function at the end of the pipeline}
}
\value{
returns NULL
}
\description{
Plot showing the number of features per cell in ascending order
}
