% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_subfunctions.R
\name{sc_atac_plot_fragments_per_cell}
\alias{sc_atac_plot_fragments_per_cell}
\title{A histogram of the log-number of fragments per cell}
\usage{
sc_atac_plot_fragments_per_cell(sce)
}
\arguments{
\item{sce}{The SingleExperimentObject produced by the sc_atac_create_sce function at the end of the pipeline}
}
\value{
returns NULL
}
\description{
A histogram of the log-number of fragments per cell
}
