% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_tss_enrichment.R
\name{sc_atac_plot_TSSE}
\alias{sc_atac_plot_TSSE}
\title{SC ATAC Plot TSS Enrichment}
\usage{
sc_atac_plot_TSSE(
  bam,
  outdir,
  txs = NULL,
  seqinfo = NULL,
  gff = NULL,
  TxDb = NULL,
  seqnames = NULL
)
}
\arguments{
\item{bam}{string path to input bam file}

\item{outdir}{string path to output directory}

\item{txs}{GRanges transcript object. Either one of `txs` and `seqinfo`, `gff` or `TxDb` file must be non null}

\item{seqinfo}{SeqInfo object containing sequence lengths. Either one of `txs` and `seqinfo`, `gff` or `TxDb` file must be non null}

\item{gff}{optional string to gff file path. Either one of `txs` and `seqinfo`, `gff` or `TxDb` file must be non null}

\item{TxDb}{TxDb type object containing sequence length data. Either one of `txs` and `seqinfo`, `gff` or `TxDb` file must be non null}

\item{seqnames}{optional character vector of sequence names to subset the TxDb or GFF file}
}
\description{
Generate a TSS enrichment plot displaying TSS scores aggregated over reads centered on the TSSs.
}
\details{
The TSS enrichment calculation is a signal to noise calculation. 
The reads around a reference set of TSSs are collected to form an aggregate distribution of reads centered on the 
TSSs and extending to 2000 bp in either direction (for a total of 4000bp). This distribution is then normalized by
taking the average read depth in the 100 bps at each of the end flanks of the distribution (for a total of 200bp of 
averaged data) and calculating a fold change at each position over that average read depth. This means that the flanks 
should start at 1, and if there is high read signal at transcription start sites (highly open regions of the genome) 
there should be an increase in signal up to a peak in the middle. We take the signal value at the center of the 
distribution after this normalization as our TSS enrichment metric.
}
