% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_remove_duplicates.R
\name{sc_atac_remove_duplicates}
\alias{sc_atac_remove_duplicates}
\title{sc_atac_remove_duplicates()}
\usage{
sc_atac_remove_duplicates(inbam, output_folder = "")
}
\value{

}
\description{
sc_atac_remove_duplicates()
}
\examples{
\dontrun{


}

}
