% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_scPipeCPP.R
\name{sc_celseq2_simulator}
\alias{sc_celseq2_simulator}
\title{sc_celseq2_simulator}
\usage{
sc_celseq2_simulator(r1fn, r2fn, annofn, bc_anno, fafn, UMI_len = 6,
  r_len = 75, frag_mean = 300, dup_mean = 5, ran_dist = "gamma",
  param = list(shape = 0.2, scale = 50), seed = NA)
}
\arguments{
\item{r1fn}{simulated read1 fastq file}

\item{r2fn}{simulated read2 fastq file}

\item{annofn}{gff3 genome annotation}

\item{bc_anno}{barcode annotation}

\item{fafn}{genome fasta file}

\item{UMI_len}{UMI length}

\item{r_len}{read length}

\item{frag_mean}{the average fragment length, for CEL-seq2 it is usually 200~500}

\item{dup_mean}{the average PCR duplicate number}

\item{ran_dist}{the distribution used to generate random gene count}

\item{param}{a list of parameter of random distribution}

\item{seed}{random seed, used for reproducing data}
}
\description{
sc_celseq2_simulator
}
\examples{
#TODO
}
