% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{sc_cluster}
\alias{sc_cluster}
\title{clustering and visualization based on PCA, diffusion map and tsne}
\usage{
sc_cluster(scd, k)
}
\arguments{
\item{scd}{an SCData object, contains gene counting matrix}

\item{k}{the number of top PCs in PCA to keep for following analysis}
}
\value{
an SCData object contains \code{reducedExprDimension}
}
\description{
TODO
}
\examples{
TODO

}

