% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_scPipeCPP.R
\name{sc_detect_bc}
\alias{sc_detect_bc}
\title{sc_detect_bc}
\usage{
sc_detect_bc(infq, outcsv, surfix = "CELL_", bc_len, max_reads = 1e+06,
  min_count = 10, max_mismatch = 1)
}
\arguments{
\item{infq}{input fastq file, shoule be the output file of \code{sc_trim_barcode}}

\item{outcsv}{output barcode annotation}

\item{surfix}{the surfix of cell name, default to be `CELL_`, the cell name will be CELL_001, CELL_002 accordingly}

\item{bc_len}{the length of cell barcode, should be consistent with bl1+bl2 in \code{sc_trim_barcode}}

\item{max_reads}{the maximum of reads processed, default is 1000,000, set to "all" to process all reads (may spend more time)}

\item{min_count}{minimum counts to keep, barcode will be discarded if it has lower count. default is 10. this should be set according to \code{max_reads}}

\item{max_mismatch}{the maximum mismatch allowed, barcodes within this number will be considered as sequence error and merged, default is 1.}
}
\description{
detect cell barcode and generate the barcode annotation
}
\examples{
\dontrun{
# `sc_detect_bc`` should run before `sc_demultiplex` for Drop-seq or 10X protocols
sc_detect_bc("input.fastq","output.cell_index.csv",bc_len=8)
sc_demultiplex(...,"output.cell_index.csv")
}


}
