% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{sc_dim_reduction}
\alias{sc_dim_reduction}
\title{dimensionality reduction based on PCA, diffusion map and tsne}
\usage{
sc_dim_reduction(scd, k = 20, n = 2)
}
\arguments{
\item{scd}{an SCData object which contains gene counting matrix or
a gene counting matrix}

\item{k}{the number of top PCs in PCA to keep for following analysis}

\item{n}{the number of dimensions after}
}
\value{
an SCData object contains \code{reducedExprDimension} or
a matrix contains t-SNE output
}
\description{
TODO
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "ensembl_gene_id")
scd = calQCMetrics(scd)
scd = detect_outlier(scd)
scd_afterqc = remove_outliers(scd)
scd_afterqc = sc_dim_reduction(scd_afterqc,n=2)
plot(DimReducedExpr(scd)$Dim1,DimReducedExpr(scd)$Dim2)

}
