% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_scPipeCPP.R
\name{sc_exon_mapping}
\alias{sc_exon_mapping}
\title{sc_exon_mapping}
\usage{
sc_exon_mapping(inbam, outbam, annofn, bam_tags = list(am = "YE", ge = "GE",
  bc = "BC", mb = "OX"), bc_len = 8, UMI_len = 6, stnd = TRUE,
  fix_chr = FALSE)
}
\arguments{
\item{inbam}{input aligned bam file}

\item{outbam}{output bam filename}

\item{annofn}{single string or vector of gff3 annotation filenames}

\item{bam_tags}{list defining BAM tags where mapping information is
stored.
\itemize{
  \item "am": mapping status tag
  \item "ge": gene id
  \item "bc": cell barcode tag
  \item "mb": molecular barcode tag
}}

\item{bc_len}{total barcode length}

\item{UMI_len}{UMI length}

\item{stnd}{TRUE to perform strand specific mapping. (default: TRUE)}

\item{fix_chr}{TRUE to add `chr` to chromosome names, MT to chrM. (default: FALSE)}
}
\value{
generates a bam file with exons assigned
}
\description{
Map aligned reads to exon annotation.
The result will be written into optional fields in bam file with different
tags. Following this link for more information regarding to bam file format:
http://samtools.github.io/hts-specs
}
\examples{
data_dir="celseq2_demo"
ERCCanno_fn = system.file("extdata", "ERCC92_anno.gff3",
    package = "scPipe")
\dontrun{
# for the complete workflow, refer to the vignettes
...
sc_exon_mapping(file.path(data_dir, "out.aln.bam"),
                file.path(data_dir, "out.map.bam"),
                ERCCanno_fn)
...
}

}
