% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warpper_scPipeCPP.R
\name{sc_gene_counting}
\alias{sc_gene_counting}
\title{Title sc_gene_counting
merge UMI and generate gene counting matrix}
\usage{
sc_gene_counting(outdir, bc_anno, UMI_cor = 1, gene_fl = FALSE)
}
\arguments{
\item{outdir}{output folder that contains \code{sc_demultiplex} output}

\item{bc_anno}{barcode annotation, first column is cell id, second column is cell barcode sequence}

\item{UMI_cor}{correct UMI sequence error: 0 means no correction, 1 means simple correction and merge UMI with distance 1.}

\item{gene_fl}{whether to remove low abundant gene count. low abundant is defined as only one copy of one UMI for this gene}
}
\description{
Title sc_gene_counting
merge UMI and generate gene counting matrix
}
\examples{
#TODO
}
