% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_tfidf.R
\name{sc_get_umap_data}
\alias{sc_get_umap_data}
\title{Generates UMAP data from sce object}
\usage{
sc_get_umap_data(sce, n_neighbours = 30)
}
\arguments{
\item{sce}{The SingleCellExperiment object}

\item{n_neighbours}{No. of neighbours for KNN}
}
\value{
A dataframe containing the UMAP dimensions, as well as all the colData of the sce object for each cell
}
\description{
Produces a DataFrame containing the UMAP dimensions, as well as all the colData of the sce object for each cell
}
