% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_tfidf.R
\name{sc_interactive_umap_plot}
\alias{sc_interactive_umap_plot}
\title{Produces an interactive UMAP plot via Shiny}
\usage{
sc_interactive_umap_plot(sce)
}
\arguments{
\item{sce}{The SingleCellExperiment object}
}
\value{
A shiny object which represents the app. Printing the object or passing it to `shiny::runApp(...)` will run the app.
}
\description{
Can colour the UMAP by any of the colData columns in the SCE object
}
