% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_multiomic_integration.R
\name{sc_multiomic_integration}
\alias{sc_multiomic_integration}
\title{Inegrating modalities from multi-omic data using barcode matching}
\usage{
sc_multiomic_integration(sce.atac, sce.rna, output_folder = NULL)
}
\arguments{
\item{sce.atac}{SCE object pre-processed for scATAC-Seq data}

\item{sce.rna}{SCE object pre-processed for scRNA-Seq data}

\item{output_folder}{The path of the output folder}

\item{barcode.match.file}{A .csv file that contains the barcode matching form each of the modalities; currently supporting only RNA and ATAC}
}
\description{
Using a barcode file match modalities of multi-omic datasets
}
\examples{
\dontrun{
sc_multiomic_integration(sce.atac = sceatac, sce.rna = scerna, barcode.match.file = barcode_file) 
}

}
