% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_sample_data.R
\docType{data}
\name{sc_sample_data}
\alias{sc_sample_data}
\title{a small sample scRNA-seq counts dataset to 
demonstrate capabilities of scPipe}
\format{a matrix instance, 1 row per gene.}
\source{
Christine Biben, WEHI
}
\usage{
sc_sample_data
}
\value{
NULL, but makes aavailable a matrix of count data
}
\description{
This data set contains counts for high variable genes for 100 cells. The cells have different cell types.
}
\author{
Luyi Tian
}
