% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_sample_qc.R
\docType{data}
\name{sc_sample_qc}
\alias{sc_sample_qc}
\title{quality control information for a small sample scRNA-seq dataset to 
demonstrate capabilities of scPipe}
\format{a data.frame instance, 1 row per cell.}
\source{
Christin Biben, WEHI
}
\usage{
sc_sample_qc
}
\value{
NULL, but makes a data frame with cell quality control data.frame
}
\description{
This data.frame contains cell quality control information for the 100 
cells.
}
\author{
Luyi Tian
}
