% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{scDimReduction}
\alias{scDimReduction}
\title{dimensionality reduction based on PCA, diffusion map and tsne}
\usage{
scDimReduction(scd, k = 20, n = 2)
}
\arguments{
\item{scd}{an SCData object which contains gene counting matrix or
a gene counting matrix}

\item{k}{the number of top PCs in PCA to keep for following analysis}

\item{n}{the number of dimensions after}
}
\value{
an SCData object contains \code{reducedExprDimension} or
a matrix contains t-SNE output
}
\description{
TODO
}
\examples{
#TODO

}
