% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{scPipe}
\alias{scPipe}
\alias{scPipe-package}
\alias{scPipe-package}
\title{scPipe - single cell RNA-seq pipeline}
\description{
It demultiplex, it quality control, it plots:
}
\author{
Luyi Tian <tian.l@wehi.edu.au>; Shian Su <su.s@wehi.edu.au>
}
