% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_integrate.R
\name{TF.IDF.custom}
\alias{TF.IDF.custom}
\title{Returns the TF-IDF normalised version of a binary matrix}
\usage{
TF.IDF.custom(binary.mat, verbose = TRUE)
}
\arguments{
\item{binary.mat}{The binary matrix}

\item{verbose}{boolean flag to print status messages}
}
\value{
Returns the TF-IDF normalised version of a binary matrix
}
\description{
Returns the TF-IDF normalised version of a binary matrix
}
