% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCData_methods.R
\docType{methods}
\name{UMI_dup_info.sce}
\alias{UMI_dup_info.sce}
\alias{UMI_dup_info,SingleCellExperiment-method}
\alias{UMI_dup_info}
\alias{UMI_dup_info<-,SingleCellExperiment-method}
\alias{UMI_dup_info}
\title{Get or set UMI duplication results in a SingleCellExperiment object}
\usage{
UMI_dup_info.sce(object)

\S4method{UMI_dup_info}{SingleCellExperiment}(object)

\S4method{UMI_dup_info}{SingleCellExperiment}(object) <- value
}
\arguments{
\item{object}{An \code{SingleCellExperiment} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
A DataFrame of UMI duplication results.
}
\description{
Get or set UMI duplication results in a SingleCellExperiment object
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
QCMetrics(scd)

}
\author{
Luyi Tian
}
