% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_sample_qc.R
\docType{data}
\name{UMI_duplication}
\alias{UMI_duplication}
\title{UMI dulication statistics for a small sample scRNA-seq dataset to 
demonstrate capabilities of scPipe}
\format{a data.frame instance.}
\source{
Christin Biben, WEHI
}
\usage{
sc_sample_qc
}
\value{
NULL, but makes a data frame with UMI dulication statistics
}
\description{
This data.frame contains UMI dulication statistics, where the
first column is the number of duplication, and the second column is the
count of UMIs.
}
\author{
Luyi Tian
}
