% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{calculate_QC_metrics}
\alias{calculate_QC_metrics}
\title{get QC metrics using gene counting matrix}
\usage{
calculate_QC_metrics(scd)
}
\arguments{
\item{scd}{an SCData object containing count}
}
\value{
no return
}
\description{
get QC metrics using gene counting matrix
}
\details{
get QC metrics using gene counting matrix
the QC statistics added are
`number_of_genes`: number of genes detected
`total_count_per_cell`: sum of read number after UMI dedup
`non_mt_percent`: 1- percent of mitochondrial gene counts
mitochondrial genes are retrived by GO term GO:0005739
`non_ERCC_percent`: ratio of exon counts to ERCC counts
`non_ribo_percent`: 1- percent of ribosomal gene counts
ribosomal genes are retrived by GO term GO:0005840
}
\examples{
TODO

}

