% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_methods.R
\name{convert_geneid}
\alias{convert_geneid}
\title{convert the gene ids of a SingleCellExperiment object}
\usage{
convert_geneid(sce, returns = "external_gene_name", all = TRUE)
}
\arguments{
\item{sce}{an SingleCellExperiment object}

\item{returns}{the gene id which is set as return. default to be `external_gene_name`
A possible list of attributes can be retrieved using the
function \code{listAttributes} from \code{biomaRt} package. the commonly used
id types are `external_gene_name`, `ensembl_gene_id` or `entrezgene`.}

\item{all}{logic. for genes that cannot covert to new gene id, keep them with the old
id or delete them. the default is keep them.}
}
\value{
sce with converted id
}
\description{
convert the gene ids of a SingleCellExperiment object
}
\details{
convert the gene id of all datas in SCData object
}
\examples{
# the gene id in example data are `external_gene_name`, the following example will convert it to `entrezgene`.
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
scd = convert_geneid(scd, return="entrezgene")

}
