% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCData_methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts.SCData}
\alias{counts,SCData-method}
\alias{counts<-,SCData,matrix-method}
\alias{counts,SCData-method}
\alias{counts}
\title{Accessors for the 'counts' element of an SCData object.}
\usage{
\S4method{counts}{SCData}(object)

\S4method{counts}{SCData,matrix}(object)<-value

\S4method{counts}{SCData}(object)

\S4method{counts}{SCData,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCData} object.}

\item{value}{an integer matrix}
}
\description{
The counts element holds the count data as a matrix of non-negative integer
count values, one row for each feature (gene, exon, region, etc), and one
column for each cell. It is an element of the assayData slot of the SCData
object.
}
\author{
Luyi Tian
}
