% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{cpm}
\alias{cpm}
\alias{cpm<-}
\alias{cpm}
\alias{cpm}
\alias{cpmSCData}
\alias{cpm,SCData-method}
\alias{cpm<-,SCData,matrix-method}
\alias{cpm}
\alias{cpm,SCData-method}
\alias{cpm<-}
\alias{cpm<-,SCData,matrix-method}
\title{Accessors for the 'cpm' (counts per million) element of an SCData object.}
\usage{
cpm(object)

cpm(object) <- value

\S4method{cpm}{SCData}(object)

\S4method{cpm}{SCData,matrix}(object)<-value

\S4method{cpm}{SCData}(object)

\S4method{cpm}{SCData,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCData} object.}

\item{value}{a matrix of class \code{"numeric"}}
}
\value{
a matrix of counts-per-million values
}
\description{
The \code{cpm} element of the arrayData slot in an SCData object holds
a matrix containing counts-per-million values. It has the same dimensions
as the 'exprs' and 'counts' elements, which hold the transformed expression
data and count data, respectively.
}
\author{
Luyi Tian
}
