% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{detect_outlier}
\alias{detect_outlier}
\title{Detect outliers based on QC metrics}
\usage{
detect_outlier(scd, comp = 1, plot = TRUE, type = c("both", "low",
  "high"), conf = c(0.9, 0.99))
}
\arguments{
\item{scd}{an SCData object containing expression values and
QC metrics.}

\item{comp}{the number of component used in GMM. can be one or two.}

\item{plot}{logical, whether to generate a pairwise plot after}

\item{type}{only looking at low quality cells (`low`) or possible doublets (`high`)
or both (`both`)}

\item{conf}{confidence interval for linear regression at lower and upper tails.
Usually this is smaller for lower tail because we hope to pick out more
low quality cells than doublets.}
}
\value{
an updated SCData object with an outlier column in \code{QualityControlInfo}
}
\description{
Detect outliers based on QC metrics
}
\details{
detect outlier using Mahalanobis distances TODO
}
\examples{
TODO

}

