% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{DimRd_expr}
\alias{DimRd_expr}
\alias{DimRd_expr<-}
\alias{DimRd_expr}
\alias{DimRd_expr.SCData}
\alias{DimRd_expr}
\alias{DimRd_expr<-}
\alias{DimRd_expr}
\title{Get or set \code{reducedExprDimension} from an SCData object}
\usage{
DimRd_expr(object)

DimRd_expr(object) <- value

DimRd_expr.SCData(object)

\S4method{DimRd_expr}{SCData}(object)

\S4method{DimRd_expr}{SCData}(object) <- value
}
\arguments{
\item{object}{An \code{\link{SCData}} object.}
}
\value{
A matrix of reduced-dimension coordinates for gene
expression of single cell.
}
\description{
Get or set \code{reducedExprDimension} from an SCData object
}
\examples{
#TODO

}
\author{
Luyi Tian
}
