% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{DimReducedExpr}
\alias{DimReducedExpr}
\alias{DimReducedExpr<-}
\alias{DimReducedExpr,SCData-method}
\alias{DimReducedExpr}
\alias{DimReducedExpr<-,SCData-method}
\alias{DimReducedExpr}
\title{Get or set \code{reducedExprDimension} from an SCData object}
\usage{
DimReducedExpr(object)

DimReducedExpr(object) <- value

\S4method{DimReducedExpr}{SCData}(object)

\S4method{DimReducedExpr}{SCData}(object) <- value
}
\arguments{
\item{object}{An \code{\link{SCData}} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
A matrix of reduced-dimension coordinates for gene
expression of single cell.
}
\description{
Get or set \code{reducedExprDimension} from an SCData object
}
\author{
Luyi Tian
}
