% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{FACSData}
\alias{FACSData}
\alias{FACSData<-}
\alias{FACSData}
\alias{FACSData.SCData}
\alias{FACSData,SCData-method}
\alias{FACSData}
\alias{FACSData<-,SCData-method}
\alias{FACSData}
\title{Get or set FACs data for an SCData object}
\usage{
FACSData(object)

FACSData(object) <- value

FACSData.SCData(object)

\S4method{FACSData}{SCData}(object)

\S4method{FACSData}{SCData}(object) <- value
}
\arguments{
\item{object}{An \code{\link{SCData}} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
A \code{AnnotatedDataFrame} of FACS data.
}
\description{
Get or set FACs data for an SCData object
}
\author{
Luyi Tian
}
