% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{gene_id_type}
\alias{gene_id_type}
\alias{gene_id_type<-}
\alias{gene_id_type}
\alias{gene_id_type.SCData}
\alias{gene_id_type}
\alias{gene_id_type<-}
\alias{gene_id_type}
\title{Get or set \code{gene_id_type} from an SCData object}
\usage{
gene_id_type(object)

gene_id_type(object) <- value

gene_id_type.SCData(object)

\S4method{gene_id_type}{SCData}(object)

\S4method{gene_id_type}{SCData}(object) <- value
}
\arguments{
\item{object}{An \code{\link{SCData}} object.}
}
\value{
gene id type string
}
\description{
Get or set \code{gene_id_type} from an SCData object
}
\author{
Luyi Tian
}
