% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCData_methods.R
\name{mergeSCData}
\alias{mergeSCData}
\title{merge multiple SCData object}
\usage{
mergeSCData(..., all = TRUE, batch = NULL)
}
\arguments{
\item{...}{multiple SCDatas. they shold have the same value for class attribute.}

\item{all}{only contains interset for features or union.}

\item{(optional)}{batch information}
}
\description{
merge multiple SCData object
}
