% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCData_methods.R
\name{newSCData}
\alias{newSCData}
\title{Create a new SCData object.}
\usage{
newSCData(exprsData = NULL, countData = NULL, tpmData = NULL,
  fpkmData = NULL, cpmData = NULL, phenoData = NULL, FACSData = NULL,
  featureData = NULL, experimentData = NULL, logged = FALSE,
  logExprsOffset = 1, reducedExprDimension = NULL,
  reducedFACSDimension = NULL, onesense = NULL, QualityControlInfo = NULL,
  useForExprs = c("exprs", "tpm", "cpm", "counts", "fpkm"))
}
\value{
a new SCData object
}
\description{
SCData extends Bioconductor's ExpressionSet class, and the same basic interface is
supported. \code{newSCData()} expects a matrix of expression values as its first
argument, with rows as features (usually genes) and columns as cells.
Per-feature and per-cell metadata can be supplied with the featureData and
phenoData arguments, respectively. Use of these optional arguments is
strongly encouraged. The SCData also includes a slot 'counts' to store an
object containing raw count data.
}
\details{
TODO
}
\examples{
TODO

}

