% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCData_methods.R
\docType{methods}
\name{organism.SCData}
\alias{organism.SCData}
\alias{organism,SCData-method}
\alias{organism}
\alias{organism<-,SCData-method}
\alias{organism}
\title{Get or set \code{organism} from an SCData object}
\usage{
organism.SCData(object)

\S4method{organism}{SCData}(object)

\S4method{organism}{SCData}(object) <- value
}
\arguments{
\item{object}{An \code{\link{SCData}} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
organism string
}
\description{
Get or set \code{organism} from an SCData object
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
organism(scd)

}
\author{
Luyi Tian
}
