% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{organism}
\alias{organism}
\alias{organism<-}
\alias{organism}
\alias{organism.SCData}
\alias{organism,SCData-method}
\alias{organism}
\alias{organism<-,SCData-method}
\alias{organism}
\title{Get or set \code{organism} from an SCData object}
\usage{
organism(object)

organism(object) <- value

organism.SCData(object)

\S4method{organism}{SCData}(object)

\S4method{organism}{SCData}(object) <- value
}
\arguments{
\item{object}{An \code{\link{SCData}} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
organism string
}
\description{
Get or set \code{organism} from an SCData object
}
\author{
Luyi Tian
}
