% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_demultiplex}
\alias{plot_demultiplex}
\title{plot_demultiplex}
\usage{
plot_demultiplex(sce)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object}
}
\value{
a pie chart
}
\description{
plot cell barcode demultiplex result for the experiment
it will include the percentage of reads that match the cell barcode.
For the reads that does not match any barcode, it will gives the percentage of
reads aligned and mapped to exon. If there is high proportion of reads
that do not match any barcode but mapped to exon, it might indicatea a failure in cell
barcode demultiplexing
}
\examples{
# TODO
}
