% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_mapping}
\alias{plot_mapping}
\title{plot mapping statistics for SCData object.}
\usage{
plot_mapping(scd, sel_col = NULL, percentage = FALSE, dataname = "")
}
\arguments{
\item{scd}{an SCData object}

\item{sel_col}{a vector of column names, indicate the columns to use for plot.
by default it will be the mapping result.}

\item{percentage}{logic. whether to convert the number of reads to percentage}

\item{dataname}{the name of this dataset which appears in the plot title}
}
\description{
plot mapping statistics for SCData object.
}
