% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotQC_pair}
\alias{plotQC_pair}
\title{plot QC statistics for SCData object}
\usage{
plotQC_pair(scd, sel_col = NULL)
}
\arguments{
\item{scd}{an SCData object}

\item{sel_col}{a vector of column names, indicate the columns to use for plot.
by default it will be the statistics generated by `calculate_QC_metrics()`}
}
\description{
plot QC statistics for SCData object
}
