% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotQC_pair}
\alias{plotQC_pair}
\title{plot QC statistics for SCData object}
\usage{
plotQC_pair(scd, sel_col = NULL)
}
\arguments{
\item{scd}{an SCData object}

\item{sel_col}{a vector of column names, indicate the columns to use for plot.
by default it will be the statistics generated by `calQCMetrics()`}
}
\value{
a ggplot2 object
}
\description{
plot QC statistics for SCData object
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
scd = calQCMetrics(scd)
scd = detect_outlier(scd)
plotQC_pair(scd)

}
