% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCData_methods.R
\docType{methods}
\name{QC_metrics.sce}
\alias{QC_metrics.sce}
\alias{QC_metrics,SingleCellExperiment-method}
\alias{QC_metrics}
\alias{QC_metrics<-,SingleCellExperiment-method}
\alias{QC_metrics}
\title{Get or set quality control metrics in a SingleCellExperiment object}
\usage{
QC_metrics.sce(object)

\S4method{QC_metrics}{SingleCellExperiment}(object)

\S4method{QC_metrics}{SingleCellExperiment}(object) <- value
}
\arguments{
\item{object}{An \code{SingleCellExperiment} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
A DataFrame of quality control metrics.
}
\description{
Get or set quality control metrics in a SingleCellExperiment object
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
QCMetrics(scd)

}
\author{
Luyi Tian
}
