% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCE_methods.R
\docType{methods}
\name{QC_metrics}
\alias{QC_metrics}
\alias{QC_metrics}
\alias{QC_metrics<-}
\alias{QC_metrics.sce}
\alias{QC_metrics,SingleCellExperiment-method}
\alias{QC_metrics}
\alias{QC_metrics<-,SingleCellExperiment-method}
\alias{QC_metrics}
\title{Get or set quality control metrics in a SingleCellExperiment object}
\usage{
QC_metrics(object)

QC_metrics(object) <- value

QC_metrics.sce(object)

\S4method{QC_metrics}{SingleCellExperiment}(object)

\S4method{QC_metrics}{SingleCellExperiment}(object) <- value
}
\arguments{
\item{object}{A \code{\link{SingleCellExperiment}} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
a dataframe of quality control matrics

A DataFrame of quality control metrics.
}
\description{
Get or set quality control metrics in a SingleCellExperiment object
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
sce = SingleCellExperiment(assays = list(counts = as.matrix(sc_sample_data)))
QC_metrics(sce) = sc_sample_qc

head(QC_metrics(sce))

}
\author{
Luyi Tian
}
