% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCData_methods.R
\docType{methods}
\name{QCMetrics}
\alias{QCMetrics}
\alias{QCMetrics<-}
\alias{QCMetrics.SCData}
\alias{QCMetrics,SCData-method}
\alias{QCMetrics}
\alias{QCMetrics<-,SCData-method}
\alias{QCMetrics}
\title{Get or set quality control metrics from an SCData object}
\usage{
QCMetrics(object)

QCMetrics(object) <- value

QCMetrics.SCData(object)

\S4method{QCMetrics}{SCData}(object)

\S4method{QCMetrics}{SCData}(object) <- value
}
\arguments{
\item{object}{An \code{\link{SCData}} object.}

\item{value}{Value to be assigned to corresponding object.}
}
\value{
A \code{AnnotatedDataFrame} of quality control metrics.
}
\description{
Get or set quality control metrics from an SCData object
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "external_gene_name")
QCMetrics(scd)

}
\author{
Luyi Tian
}
