% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{remove_outliers}
\alias{remove_outliers}
\title{remove outliers for SCData}
\usage{
remove_outliers(scd)
}
\arguments{
\item{scd}{an SCData object}
}
\description{
remove outliers for SCData
}
\examples{
data("sc_sample_data")
data("sc_sample_qc")
QualityControlInfo = new("AnnotatedDataFrame", data = as.data.frame(sc_sample_qc))
scd = newSCData(countData = as.matrix(sc_sample_data),
               QualityControlInfo = QualityControlInfo,
               useForExprs = "counts",
               organism = "mmusculus_gene_ensembl",
               gene_id_type = "ensembl_gene_id")
scd = calQCMetrics(scd)
scd = detect_outlier(scd)
scd_afterqc = remove_outliers(scd)

}
