% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_aligning.R
\name{sc_atac_aligning}
\alias{sc_atac_aligning}
\title{sc_atac_align()}
\usage{
sc_atac_aligning(
  ref,
  readFile1,
  readFile2 = NULL,
  readDir = NULL,
  output_folder = ""
)
}
\value{

}
\description{
sc_atac_align()
}
\examples{
\dontrun{


}

}
