% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_bam_tagging.R
\name{sc_atac_bam_tagging}
\alias{sc_atac_bam_tagging}
\title{sc_atac_bam_tagging()}
\usage{
sc_atac_bam_tagging(
  inbam,
  outbam = "",
  bam_tags = list(bc = "CB", mb = "OX"),
  nthreads = 1
)
}
\value{

}
\description{
sc_atac_bam_tagging()
}
\examples{
\dontrun{


}

}
