% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_cell_calling.R
\name{sc_atac_cell_calling}
\alias{sc_atac_cell_calling}
\title{sc_atac_bam_tagging()}
\usage{
sc_atac_cell_calling(
  mat,
  cell_calling,
  output_folder,
  genome_size = NULL,
  qc_per_bc_file = NULL,
  lower = NULL
)
}
\value{

}
\description{
sc_atac_bam_tagging()
}
\examples{
\dontrun{


}

}
