% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_create_fragments.R
\name{sc_atac_create_fragments}
\alias{sc_atac_create_fragments}
\title{generating the popular fragments for scATAC-Seq data using sinto}
\usage{
sc_atac_create_fragments(inbam, output_folder = "")
}
\value{

}
\description{
Takes in a tagged and sorted BAM file and outputs the associated fragments in a .bed file
}
\examples{
\dontrun{


}

}
