% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_create_report.R
\name{sc_atac_create_report}
\alias{sc_atac_create_report}
\title{HTML report generation}
\usage{
sc_atac_create_report(
  input_folder,
  feature_file,
  output_folder = NULL,
  organism = NULL,
  sample_name = NULL,
  feature_type = NULL,
  n_barcode_subset = NULL,
  tss_file = NULL,
  promoter_file = NULL,
  enhancer_file = NULL
)
}
\arguments{
\item{input_folder}{The path of the folder produced by the pipeline}

\item{feature_file}{is the peak or genome_bin file that was generated via ScPipe workflow or outside the workflow (.bed format)}

\item{output_folder}{The path of the output folder to store the HTML report in}

\item{organism}{A string indicating the name of the organism being analysed}

\item{sample_name}{A string indicating the name of the sample}

\item{feature_type}{A string indicating the type of the feature (`genome_bin` or `peak`)}

\item{n_barcode_subset}{if you require only to visualise stats for a sample of barcodes to improve processing time (integer)}

\item{tss_file}{bed file of the TSS regions}

\item{promoter_file}{bed file of the promoter regions}

\item{enhancer_file}{bed file fo the enhancer regions}
}
\description{
Generates a HTML report using the output folder produced by the pipeline
}
