% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_create_report.R
\name{sc_atac_create_report}
\alias{sc_atac_create_report}
\title{sc_atac_create_report()}
\usage{
sc_atac_create_report(
  input_folder,
  output_folder,
  sample_name = NULL,
  organism = NULL,
  feature_type = NULL
)
}
\value{

}
\description{
sc_atac_create_report()
}
\examples{
\dontrun{


}

}
