% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_create_report.R
\name{sc_atac_create_report}
\alias{sc_atac_create_report}
\title{HTML report generation}
\usage{
sc_atac_create_report(
  input_folder,
  output_folder = NULL,
  organism = NULL,
  sample_name = NULL,
  feature_type = NULL,
  n_barcode_subset = 500
)
}
\arguments{
\item{input_folder}{The path of the folder produced by the pipeline}

\item{output_folder}{The path of the output folder to store the HTML report in}

\item{organism}{A string indicating the name of the organism being analysed}

\item{sample_name}{A string indicating the name of the sample}

\item{feature_type}{A string indicating the type of the feature (`genome_bin` or `peak`)}

\item{n_barcode_subset}{if you require only to visualise stats for a sample of barcodes to improve processing time (integer)}
}
\value{
the path of the output file
}
\description{
Generates a HTML report using the output folder produced by the pipeline
}
