% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_create_sce.R
\name{sc_atac_create_sce}
\alias{sc_atac_create_sce}
\title{sc_atac_create_sce()}
\usage{
sc_atac_create_sce(
  input_folder = NULL,
  organism = NULL,
  feature_type = NULL,
  pheno_data = NULL,
  report = FALSE
)
}
\value{

}
\description{
sc_atac_create_sce()
}
\examples{
\dontrun{


}

}
