% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_create_sce.R
\name{sc_atac_create_sce}
\alias{sc_atac_create_sce}
\title{sc_atac_create_sce()}
\usage{
sc_atac_create_sce(
  input_folder = NULL,
  organism = NULL,
  sample_name = NULL,
  feature_type = NULL,
  pheno_data = NULL,
  report = FALSE
)
}
\arguments{
\item{input_folder}{The output folder produced by the pipeline}

\item{organism}{The type of the organism}

\item{sample_name}{The name of the sample}

\item{feature_type}{The type of the feature}

\item{pheno_data}{The pheno data}

\item{report}{Whether or not a HTML report should be produced}
}
\value{
a SingleCellExperiment object created from the scATAC-Seq data provided
}
\description{
sc_atac_create_sce()
}
\examples{
\dontrun{
sc_atac_create_sce(
   input_folder = input_folder,
   organism = "hg38",
   feature_type = "peak",
   report = TRUE)
}    

}
