% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_feature_counting.R
\name{sc_atac_feature_counting}
\alias{sc_atac_feature_counting}
\title{TODO: documentation}
\usage{
sc_atac_feature_counting(
  insortedbam,
  feature_input,
  bam_tags = list(bc = "CB", mb = "OX"),
  feature_type = "peak",
  bin_size = NULL,
  yieldsize = 1e+06,
  mapq = 0,
  blacklist = NULL,
  output_folder = "",
  fix_chr = "none"
)
}
\description{
TODO: documentation
}
