% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_feature_counting.R
\name{sc_atac_generate_qc_plots}
\alias{sc_atac_generate_qc_plots}
\title{generating qc plots}
\usage{
sc_atac_generate_qc_plots(
  frags_file,
  peaks_file,
  qc_per_bc_file,
  output_folder
)
}
\arguments{
\item{frags_file}{The fragment file}

\item{peaks_file}{The peak file}

\item{output_folder}{The path of the output folder}

\item{inbam}{The input BAM file}
}
\description{
qc plots are produced using fragment file and peak file for features
}
