% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_peak_calling.R
\name{sc_atac_peak_calling}
\alias{sc_atac_peak_calling}
\title{sc_atac_call_peaks()}
\usage{
sc_atac_peak_calling(inbam, output_folder = "")
}
\value{

}
\description{
sc_atac_call_peaks()
}
\examples{
\dontrun{


}

}
