% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_peak_calling.R
\name{sc_atac_peak_calling}
\alias{sc_atac_peak_calling}
\title{sc_atac_peak_calling()}
\usage{
sc_atac_peak_calling(
  inbam,
  ref = NULL,
  genome_size = NULL,
  output_folder = NULL
)
}
\arguments{
\item{inbam}{The input tagged, sorted, duplicate-free input BAM file}

\item{ref}{The reference genome file}

\item{genome_size}{The size of the genome}

\item{output_folder}{The path of the output folder}
}
\value{
None (invisible `NULL`)
}
\description{
sc_atac_peak_calling()
}
\examples{
\dontrun{
sc_atac_peak_calling(
    inbam,
    reference) 
}
}
