% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_pipeline.R
\name{sc_atac_pipeline}
\alias{sc_atac_pipeline}
\title{A convenient function for running the whole pipeline}
\usage{
sc_atac_pipeline(
  r1,
  r2,
  barcode_fastq,
  organism,
  reference,
  feature_type,
  remove_duplicates = FALSE,
  samtools_path = NULL,
  cell_calling = FALSE,
  output_folder = NULL
)
}
\arguments{
\item{r1}{The first read fastq file}

\item{r2}{The second read fastq file}

\item{barcode_fastq}{The barcode fastq file}

\item{organism}{The name of the organism e.g. hg38}

\item{reference}{The reference genome file}

\item{feature_type}{The feature type (either `genome_bin` or `peak`)}

\item{remove_duplicates}{Whether or not to remove duplicates (samtools is required)}

\item{samtools_path}{A custom path of samtools to use for duplicate removal}

\item{cell_calling}{The desired cell calling method (either `emptydrops`, `filter`, or `cellranger`)}

\item{output_folder}{The path of the output folder}
}
\description{
A convenient function for running the whole pipeline
}
