% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atac_trim_barcode.R
\name{sc_atac_trim_barcode}
\alias{sc_atac_trim_barcode}
\title{sc_atac_trim_barcode()}
\usage{
sc_atac_trim_barcode(
  r1,
  r2,
  bc_file,
  output_folder = "",
  bc_start = -1,
  bc_length = -1,
  umi_start = 0,
  umi_length = 0,
  umi_in = "both",
  rmN = FALSE,
  rmlow = FALSE,
  min_qual = 20,
  num_below_min = 2,
  id1_st = -1,
  id1_len = -1,
  id2_st = -1,
  id2_len = -10
)
}
\value{

}
\description{
sc_atac_trim_barcode()
}
\examples{
\dontrun{


}

}
