% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_scPipeCPP.R
\name{sc_exon_mapping}
\alias{sc_exon_mapping}
\title{sc_exon_mapping}
\usage{
sc_exon_mapping(inbam, outbam, annofn, am = "YE", ge = "GE", bc = "YC",
  mb = "YM", bc_len = 8, UMI_len = 6, stnd = TRUE, fix_chr = FALSE)
}
\arguments{
\item{inbam}{input bam file. this is the output of alignment program}

\item{outbam}{output bam file with gene and barcode tag}

\item{annofn}{genome annotation, gff file. can have multiple files}

\item{am}{mapping status tag (default: YE)}

\item{ge}{gene id tag (default: GE)}

\item{bc}{cell barcode tag (default: YC)}

\item{mb}{molecular barcode tag (default: YM)}

\item{bc_len}{total barcode length}

\item{UMI_len}{UMI length}

\item{stnd}{perform strand specific mapping or not}

\item{fix_chr}{add `chr` to chromosome names, fix inconsistant names.}
}
\description{
take the alinged read and map them to exons. the result will be written into optional fields
in bam file with different tags. for more information regarding to bam file format:
http://samtools.github.io/hts-specs/SAMv1.pdf
}
\examples{
#TODO
}
