% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_integrate.R
\name{sc_mae_plot_umap}
\alias{sc_mae_plot_umap}
\title{Generates UMAP of multiomic data}
\usage{
sc_mae_plot_umap(mae, by = NULL, output_file = NULL)
}
\arguments{
\item{mae}{The MultiAssayExperiment object}

\item{by}{What to colour the points by. Needs to be in colData of all experiments.}

\item{output_file}{The path of the output file}
}
\value{
A ggplot2 object representing the UMAP plot
}
\description{
Uses feature count data from multiple experiment objects to produce UMAPs for each assay and then overlay them on the same pair of axes
}
