% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_scPipeCPP.R
\name{sc_trim_barcode}
\alias{sc_trim_barcode}
\title{sc_trim_barcode}
\usage{
sc_trim_barcode(outfq, r1, r2 = NULL, read_structure = list(bs1 = -1, bl1 =
  2, bs2 = 6, bl2 = 8, us = 0, ul = 6), filter_settings = list(rmlow = TRUE,
  rmN = TRUE, minq = 20, numbq = 2))
}
\arguments{
\item{outfq}{the output fastq file, which reformat the barcode and UMI into the read name}

\item{r1}{read one for pair-end reads. This read should contains the transcript}

\item{r2}{read two for pair-end reads. default to be `NULL` for single reads}

\item{read_structure}{a list contains read structure configuration:\itemize{
\item{bs1} is the starting position of barcode in read one, if there is no reads in read one set it to -1.
\item{bl1} is the length of barcode in read one, if there is no barcode in read one this number is used for trimming the beginning of read one.
\item{bs2} is the starting position of barcode in read two
\item{bl2} is the length of barcode in read two
\item{us} is the starting position of UMI
\item{ul} is the length of UMI
}}

\item{filter_settings}{a list contains read filter settings:}
}
\description{
reformat the fastq file and move the barcode and UMI to read names.
}
\examples{
#TODO
}
